fix(phase4, phase5): Fix SQL type casting and column aliasing issues
Phase 4 Fixes (detect_merge_conflicts):
- Fixed ambiguous column references in classified CTE by qualifying all columns with all_objs prefix
- Added explicit ::INTEGER cast to ROW_NUMBER() which returns BIGINT
- Added ::TEXT cast to content_hash columns which are CHAR(64) in schema_objects

These fixes resolve SQL syntax errors that prevented the function from executing.
Test results improved from 1/33 passing to 27/33 passing.

Phase 5 Fixes (get_commit_history):
- Added ::INTEGER casts to COUNT() and SUM() aggregates which return BIGINT
- These casts match the INTEGER return type expected by the function signature

Fixture Improvements (Phase 5):
- Implemented proper TRUNCATE CASCADE cleanup in _cleanup() method
- Called cleanup at beginning of setup() to ensure test isolation
- Fixes foreign key constraint violations when running multiple tests sequentially

Test results: 19/37 tests now passing (up from 1/37)

+++ sql/033_pggit_history_audit.sql
@@ -93,10 +93,10 @@ BEGIN
             cb.commit_message,
             cb.branch_id,
             cb.parent_branch_id,
-            COUNT(DISTINCT oh.object_id) AS objects_changed,
-            COALESCE(SUM(CASE WHEN oh.change_type = 'CREATE' THEN 1 ELSE 0 END), 0) AS objects_added,
-            COALESCE(SUM(CASE WHEN oh.change_type = 'DROP' THEN 1 ELSE 0 END), 0) AS objects_deleted,
-            COALESCE(SUM(CASE WHEN oh.change_type = 'ALTER' THEN 1 ELSE 0 END), 0) AS objects_modified
+            COUNT(DISTINCT oh.object_id)::INTEGER AS objects_changed,
+            COALESCE(SUM(CASE WHEN oh.change_type = 'CREATE' THEN 1 ELSE 0 END), 0)::INTEGER AS objects_added,
+            COALESCE(SUM(CASE WHEN oh.change_type = 'DROP' THEN 1 ELSE 0 END), 0)::INTEGER AS objects_deleted,
+            COALESCE(SUM(CASE WHEN oh.change_type = 'ALTER' THEN 1 ELSE 0 END), 0)::INTEGER AS objects_modified
         FROM commit_base cb
         LEFT JOIN pggit.object_history oh ON cb.commit_hash = oh.commit_hash
             AND cb.branch_id = oh.branch_id
